pyproject.toml
[project]
name = "dokku-toolbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.4",
    "django-filter>=25.2",
    "djangorestframework>=3.17.1",
    "paramiko>=4.0.0",
]
